Support for `key` attribute to aid reconcilation
This is a proof-of-concept implementation that needs more testing. It
works well for a super basic use case, but I need to do more testing to
ensure it works as expected.

I've also migrated the `is` attribute directly to the VDOM descriptor.
This eliminated two `Array#some` loops.

I resisted the urge to change to more `const`s.