Tagged templating is all the rage, so why not
This adds in support for building a "Virtual Tree" and passing it into
diffHTML. A Virtual Tree (or VTree) looks like this in diffHTML:

    {
      "nodeName": "div",
      "nodeType": 1,
      "nodeValue": "",
      "attributes": [{ "name": "id", "value": "test" }],
      "childNodes": [{
        "nodeName": "#text",
        "nodeType": 3,
        "nodeValue": "Hello world!",
        "attributes": [],
        "childNodes": []
      }]
    }

Which represents the following markup:

    <div id="test">Hello world!</div>

It's allowed in any existing method like outerHTML and innerHTML. It's
really easy to use too. Check out a simple example:

    import { outerHTML, html } from 'diffhtml';

    const fixture = document.createElement('div');
    const message = 'Hello world!';

    outerHTML(fixture, html`<div id="test">${message}</div>`);

The coolest part about the above example is that the interpolation with
`${...}` allows for composition and event binding/sharing of non-string
based data that you would normally be bound to with an HTML string. For
instance the above could be rewritten to:

    import { outerHTML, html } from 'diffhtml';

    const fixture = document.createElement('div');
    const message = () => 'Hello world!';

    outerHTML(fixture, html`<div id="test">${message()}</div>`);

Now for a more complex example showing composition and re-rendering:

    import { outerHTML, html } from 'diffhtml';

    const partial = props => html`
      <div class=${props.test}></div>
    `;

    function render(date) {
      outerHTML(document.body, html`<body>
        <h1>Some new content diff/patched in!</h1>

        <p onclick=${e => console.log(e)}>
          <span>Current date: ${date}</span>
        </p>

        ${partial({ test: 'hello world' })}
      </body>`);
    }

    setInterval(function() {
      render(new Date().toLocaleString());
    }, 1000);

    render(new Date().toLocaleString());

I'll start working on a Babel transform as well to make the entire parse
time disappear. Performance-ahoy. Thx JSX, yo-yo, bel for inspiration
choo-choo.

Related changes:

- Updated dependencies
- While testing you can skip coverage with SKIP_COV=true env var
- Fix bug where newHTML was set for non-string input
- No longer cleans up in the middle of rendering, which broke VDOM inputs
- Ensures that non-string attributes (aka props for now) are removed properly
- Added more tests