Refactor transaction completion + Babel polyfill
This adds in the Babel polyfill and removes all the others. I've added a
note to the README for including it in projects that need legacy
support.

I've also changed the util/render.js to be finalize.js under the node/
directory. The function name got updated to getFinalizeCallback which
more accurately summarizes what it does than completeRender.