Fixes to the tests to support various browsers
Not all browsers behaved with the tests, so I've gone through with
Browserstack and fixed the files that were not valid. Such as creating
an HTML Document without a title (apparently it's a required argument in
some browsers).

I can't get IE 11 to work just yet. I have an undocumented internal
event named `renderComplete`. I may have mentioned it on the website,
but it's definitely going away in it's current form. I'll bring it back
as middleware if anyone needs it.