No longer protect/unprotect during tree sync
Synchronization should only compare and modify the original Virtual Tree
Element to the new Virtual Tree Element. It should not be
protecting/unprotecting objects. The patch stage should do this.

I found oddness with the attribute synchronization, so I've updated the
logic to always send the old and new attributes. This avoids a DOM
attribute lookup during the patching process.