Fixes handling of interpolated attributes
The isProp regular expression was insufficient as it only checked the
preceding string, and if you had multiple values, the second preceding
string would be a single space (attribute value delimiter). Now we run
the regex starting from the end and working its way back to determine if
we're in an attribute/property.