Updates Components, ESM, and DevTools
Adds in the devtools that were promised (not fully complete yet).
Improves the components story, but also is not yet complete (hard to do
man!) I've added a ton of tests and greatly improved the strict mode
feature (lots left to do...) Greatly improves the ESM story, although it
has to use relative paths for now... sad but true. Eventually folks will
land a loader, but until then relative paths is the safest bet.