Fixes Component rendering in render-to-string
When a Component is rendered it would create a document-fragment and
diff against that on the root. This broke nested child rendering. Now
innerHTML is used instead inside of a container to fully collect any
render trees. All existing tests pass! The website is now generated with
the new render-to-string code.