Continue to improve codebase
This commit adds in a VTreeLike type which accounts for JSX and Custom
Element types. All input is now normalized through `createTree` to
ensure a pooled object represents the value.

Prior to this commit `cleanMemory` would clean up all NodeCache
associations. This is no longer the case, as it changes the robustness
of the tests.

- All VTree->NodeCache associations are cleaned up when deallocated.

- `cleanMemory()` is now `gc()`

- `createTree()` always returns a VTree now