Resolve namespace issues when creating SVG element
Prior to this commit, SVG would only work if rendered directly from the
top, like: `diff.innerHTML(document.body, '<svg></svg>');` and not when
diffing or creating within an existing document.

This corrects the behavior by determining the correct type during
synchronization, applying this VTree reference to an `svgElements` set,
and then referencing this when creating elements in the patch phase.

This should not incur any memory overhead or significant processing and
keeps our VTree shape lean, and filesize low by not including a list of
all SVG elements.