Call createTree on parser close, not open (#289)
* Call createTree on parser close, not open

In the new parser createTree was called with only a tag name and not the
attributes or childNodes. This PR fixes the new parser to wait until the
element is closed before calling createTree. This work uncovered some
inconsistencies and improved the parser by closing out the pointer and
resetting to the parent in a function.

* Update comment

* Remove unused prop

* Satisfy type check